First implementation of Metricbeat
This is the first basic implementation of MetricBeat. For more details on
what it contains, check below.

= MetricSets
The following basic MetricSet are part of this first version.

* Apache module with Apache status
* Redis module with Redis info
* MySQL module with status

Be aware that only very few metrics are collected per MetricSet, but more
could be added easily

= Data Model

The current data model of each metricset can be found in the file
beater/metricbeat.go in the header docs.

= Error Handling

It is assumed, that a module or metricset can panic. This panic is catched
to not kill all other modules. This error is reported in the log files. The failing
module / metricset will not be restarted

= Configuration

Each module and metricSet can have its own configuration. The following fields
are generic configurations per module and are always available:

* Period
* Hosts

More details can be found in etc/beat.yml

= Testing

* Golang test framework for unit tests
* Golang test framework combined with docker-compose for integration tests
* Python libbeat system test framework combined with docker-compose for system tests

All tests are intended to be run inside the docker environment.

= Vendoring

The prevent dependency conflicts between modules, each module can have its own vendor
environment. This is implemented as an example in the MySQL module.