Simplify state handling in Filebeat (#1473)
* Simplify state handling in Filebeat

Before there were 2 different channels to report state so also prospector and registrar could report the state. State is now always reported by events from the harvesters. This means that events can also be empty to only update state. This means if 3 events were processed, this does not necessarily mean that 3 events were sent, as some of the events can also be state information.

* Introduction of channel to prospector for events. This allows for actions on the event on a prospector level
* Cleanup variable naming

Part of #1022

* Prevent race in state