Making modules reusable outside of Metricbeat (#1616)
ModuleWrapper is now exported so that it can be used outside of the
package. ModuleWrapper provides a simple Start() method that outputs
to a golang channel. This makes testing of ModuleWrapper quite easy
because it has few outside dependencies.

Added a facade called ModuleRunner that provides simple Start() and
Stop() methods and publishes to a publisher.Client.

There are godoc examples demonstrating how to use both ModuleWrapper
and ModuleRunner.