Remove exporting the index pattern (#2350)
Pass as option to the script a regexp to match the Kibana dashboards that need to be exported. This script is not
exporting the index pattern anymore, as it might overwrite the generated index pattern.
The make export-dashboards can be used to export all the Kibana dashboards for a Beat.