Improve handling of messages larger than 10MB
Old implementation dropped messages larger than 10MB (hard coded),
to protect against memory DoS. However, in some cases the transaction
was still recorded, but the parsing of the headers was incomplete.

This makes the support for large messages explicit, by adding a mode
to the parser that "sees" the segments without storing them.