Fix 'index out of range' bug in Packetbeat DNS protocol plugin (#2875)
When converting resource records into MapStrs the code can run into an index out of range exception when processing OPT psuedo resource records. The code reduces the size of a slice by creating a new, smaller slice and copies the existing data to the smaller slice. But the code did not account for the fact that the length of the slice was smaller when indexing.

* Added a DNS test case for rrsToMapStrs with a OPT resource record

Fixes #2872