Increase the fields limit (#3275)
Elasticsearch has a default limit of 1000 fields starting with 5.x.
Metricbeat already defines a bit over that in it's current mapping. In most
deployments, these fields won't be actually used all at the same time, so
the effects of increasing the limit shouldn't be that large (can still affect
the cluster state processing).

Longer term we'll need a better solution for this.