Improve template testing (#3425)
Template tests broke in the past several times because we have a test template which is not updated automatically. The tests are now changed so that the libbeat template is generated and tested. This guarantees that the templates we actually use are tested. This requires for integration tests also for libbeat to be called `make update`.

* Switch to `update` also for libbeat.
* Files generated by `update` are added go `.gitignore` as these files should not be versioned.
* Add heartbeat to template tests
* Remove static test templates as not needed anymore.