Use more global paths for system tests (#3591)
This change is the first step to make it possible to execute system tests from different locations. Currently all system test execution relies on the convention that everything is under `tests/system/` and that `nosetests` is executed in this directory. Based on this it is not possible to place system tests in different location, for example inside a metricbeat module. In addition it should simplify the creation of system tests in community beats.

* Unify handling of setUpClass
* Rename SystemTest to Test to be consistent
* Rename `beat_path` variable to `test_binary` because that is what it is
* Expect full template path in template loader
* Use absolute paths for importing other python files
* Remove duplicated code from packetbeat setUp