Use Docker Healthcheck to wait for services to start (#3652)
Let Docker manager the start order of containers based on the dependencies declared
in the docker-compose configuration files. The removes the need to have custom
Docker entrypoint scripts that test services and wait for them to start before proceeding.

Added 127.0.0.1 to the ports in local.yml so that the containers cannot be connected to remotely.

Updated to Elastic stack 5.2.1 for latest.yml environment.

Use docker-compose 1.11 on Travis CI.

Remove docker-entrypoint.sh scripts.