Assert on errors in the filebeat modules (#4186)
We used to whitelist the mysql/slowlog and system/auth fileset when
checking if there were any errors while parsing. This removes all whitelisted
entries and always fails if there are error keys generated.

The mysql module was whitelisted because Mysql writes a header that doesn't match
the patterns. I solved this by excluding the header via the Filebeat config.

The auth fileset was already fixed to not generate errors, at least for the current
test set.