libbeat: add `process.GetOne` to get one process by PID
Previously, if a Beat was started as root the debug logs are flooded
with these messages every 30 seconds before a monitoring event is created:
`2018-01-08T22:40:09.884+0100    DEBUG   [processes]
process/process.go:381  Skip process pid=3: error getting process exe
for pid=3: readlink /proc/3/exe: no such file or directory`.

It happens because the current `process.Get` iterates over all PIDs of the
current processes, but it cannot access everything. As only the Beat
process is interesting to us, there is no need to try to create a
Process for every PID on the host.

I added a new function with a different algorith,
which gets a PID and fetches process data for that PID.
This way, there is no unnecessary iterating over list of PIDs.
I simply pass the PID of the running Beat process and get data for only
it. So these messages never show up.

I extracted it into a smaller function so getting a data for a PID can
be used from both process.Get and process.GetOne. I preserved the
format, level and the content of the error messages.