Do not fail recursive watches due to unaccessible subdir (#5958)
* Do not fail setting recursive watchers (#5911)

When a subdirectory of a watched dir is not accessible, due to
permissions for example, do not fail watching the whole tree.

Test by moving a directory tree into a watched dir, where the moved tree
contains a non accessible subdir.

* Remove output to terminal from test

Use testing package log for output instead of the terminal. This way
it can be controlled by the testing.v flag.