auditbeat: Warn if auditd is running (#6023)
* Auditbeat: Detect SetPID failure

Detect failures When Auditbeat is installed as audit process by setting
the PID field in the AuditStatus structure. This usually means another
process is already set as the audit process.

* Update vendored go-libaudit from v0.0.6 to master

* Auditbeat: Don't set rules if audit status is locked

The audit rules can be locked (enabled=2) so that further
changes are not possible. Skip rule configuration if this
is the case, displaying a warning message if rules are set
in the configuration.

* Auditbeat: Fail unicast configuration if audit locked

Closes #6019

* Report errors when adding rules