filebeat: make pipeline updating configurable (#6417)
New flag is introduced: `--update-pipelines`.

It is quite uncomfortable to delete Ingest pipelines manually before testing a small change. So during development no one cares if pipelines are updated by Filebeat unnecessarily. Or if someone does he/she can stick with the default behaviour by not specifying `--update-pipelines`.

It is intended to be used during module development or testing pipeline changes.
It should not be used when running Beats in production, because it's unnecessary to update pipelines every time a Beat is started.