allow forcing dynamic object type mapping (#6691)
Currently dynamic mappings are tolerant of unexpected data types.  For example, this definition:
```
- name: test
    type: object
    object_type: long
    dynamic: true
```

produces this dynamic template:
```
"test": {
  "path_match": "test.*",
  "match_mapping_type": "long",
  "mapping": {
    "type": "long"
  }
}
```

If the first value seen is a string, say indexing this document:
```
"test": {
      "somelong": 16,
      "somestring": "foo"
}
```

the resulting mapping is actually:
```
"test": {
  "dynamic": "true",
  "properties": {
    "test": {
      "properties": {
        "somelong": {
          "type": "long"
        },
        "somestring": {
          "type": "keyword",
          "ignore_above": 1024
        }
      }
    }
  }
}
```

In apm-server, data is already validated before being sent along to elasticsearch, so barring a bug, a string will not be sent where a long is required.  This change allows forcing the data for a dynamic field to be a specific type via field config, eg for:

```
- name: test
    type: object
    object_type: long
    object_type_mapping_type: "*"
    dynamic: true
```

The resulting dynamic template would be:

```
"test": {
  "path_match": "test.*",
  "match_mapping_type": "*",
  "mapping": {
    "type": "long"
  }
}
```

Any value that can't be mapped to `long` will result in an indexing error.  Continuing this example:
```
"error": {
  "root_cause": [
    {
      "type": "mapper_parsing_exception",
       "reason": "failed to parse [test.somestring]"
    }
  ]
}
```