Introduce environment variable to skip compose starting environment (#6716)
By default the golang integration tests in metricbeat use docker compose to start their own environment for testing. This is useful for quick and automated testing. But in some cases a local instance of the service is used for developing and testing a metricset. This PR introduces the `NO_COMPOSE` environment variable which can be used to skip the starting of the service with compose.

To generate data for logstash.node_stats metricset from a local LS instance, the following command can be used:

```
NO_COMPOSE=1 go test -tags=integration github.com/elastic/beats/metricbeat/module/logstash/node_stats/... -data
```