Allow autodiscover to monitor unexposed ports (#6727)
The metrics hint builder needs a slight change to allow monitoring of endpoints like `${data.host}:9090`. Currently it blocks such cases. For host network containers, there wouldnt be any port on the configuration as well. We would just spin up metricbeat modules, if port is 0 and the host hint has `data.host` on it.