Do not log errors when X-Pack Monitoring is not available (#6627)
X-Pack Monitoring feature keeps retrying to connect to Elasticsearch until either Elasticsearch is available and the Monitoring is enabled. Before each time a connection attempted failed an errors was logged. This PR changes the implementation that only an info is printed about not being able to connect on the first attempt. After that reporting will keep retrying to connect to Elasticsearch until it is successful, but no errors are logged. On success an message is logged that the connection was successful.

If for whatever reason X-Pack was available but at a later point isn't anymore, Monitoring will log errors as this is not an expected case.

Note: In Elasticsearch 6.2 Monitoring as an interval parameter for the collection. If it is set to `-1` collection is disabled but events will still be written to Elasticsearch. The issue with this in 6.2 is that no cleanup of the sent events happen. In 6.3 there is a collection config option to turn collection off even if monitoring is enabled. If collection is turned off the bulk requests are still accepted and return a 200, but that will not be persisted. The Beat on startup only checks if Monitoring is installed. If collection is enabled or not is a setting internal to monitoring and monitoring is expected to do the right thing with the events here.

This PR has been manually tested against 6.2 but no automated tests were implemented yet.