Opening X-Pack
This adds the x-pack directory and associated license files. It updates the
packaging to produce two sets of binaries. And it updates the integration
tests to be able to consume both the default artifacts and the OSS artifacts
of other Elastic Stack products.

There are some Makefile target changes for packaging.

- `make package-all` - Creates two sets of artifacts -- ones with and without x-pack content. The dev-tools/deploy script will use this target.
- `make package-oss` - Creates packages without Elastic X-Pack content.
- `make package-elastic` - Creates packages with Elastic X-Pack content.
- `make package` - Left unchanged in order to not affect libbeat users of this target.

The code generators were updated to remove the x-pack folder upon initial creation.

In the deb packages the license metadata has changed from "ASL 2.0" to "ASL-2.0" in compliance
with https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/#license-syntax.

Update Docker testing environment to support oss and non-oss

- For the ES/LS/Kibana images, download either the OSS package or the X-Pack package based on the IMAGE_FLAVOR variable (previously it was XPACK=1 or 0).
- Must explicitly enable a X-Pack trial license in the ES image for testing with ML and Alerting.
- Refactor the LS image to be closer to the elastic/logstash-docker repo. It now honors the IMAGE_FLAVOR variable.