Metricbeat: Remove duplicated filesystems in system module (#6819)
* Metricbeat: Remove duplicated filesystems in system module

Mountpoints whose device name is the same absolute path are considered
to be the same and are counted only once. The directory name used is
the shorter between them. This avoids duplication of filesystems as can
happen with bind mounts in Linux (#3384).

* Metricbeat: Ignore nodev filesystems by default

In system module, when no filesystem type is set to be ignored, it
tries to make a sane guess. In systems with /proc/filesystems file it
ignores all devices marked as `nodev`.