Auditbeat recursive file watches for Windows (#6893)
* Add support for case-insensitive text search in system-tests

Allow system tests to search for text in the beat logfile using
case-insensitive search. This is necessary to match paths in
case-insensitive file systems, where the path logged may have different
capitalisation than the one used in the system test.

* Use custom version of fsnotify with recursion support

* Added system-test for the file_integrity module

* Auditbeat: Fix deadlock in fsnotify under Windows

Under Windows, directories to be watched need to be installed after the
event consumer loop is started. Otherwise there's a chance of a
potential deadlock, as fsnotify under Windows uses a single goroutine to
deliver events and install watches. If the channel it uses to deliver
events is full, it will block and won't be able to install further
watches.

* Auditbeat: Use OS-support for recursive watches

This patch enables OS-supported recursive watching in fsnotify if it is
available. Currently only supported under Windows via our custom
fsnotify fork.

Fixes #6864