Disable fsnotify recursive test under Darwin (#6962)
A race has been discovered in the kqueue implementation of fsnotify
which causes installation of a watch to a subdirectory of an already
watched directory to fail occasionally.

Disabled the test because Auditbeat doesn't use fsnotify under Darwin,
so there is no point on testing it anyway.

From an initial investigation it resulted that  sometimes a directory
creation event from the kernel is received twice. In Auditbeat, we react
to the first one by adding a watch to the directory. Once the second
event arrives, fsnotify overwrites the watch with an internal one, as it
needs to install a watch on every file and directory it monitors.