Switch to debug logs for missing fields (#6928)
Currently every time a required field is missing in a metricset a log entry is created. This can create a lot of log messages. This is now a temporary fix to only log missing fields on debug level and log all missing fields in one message instead of having one for each field.

The logs still lack context as the schema does not know from which metricset it is called. Further improvements to follow.

The `ErrorDebug` method was removed as it was not used.