Use minikube on travis for K8S integration tests (#7096)
This change introduces minikube on travis for tests. This will allow us to perform integration tests for k8s related code. At the moment we only test that deploy manifests are valid, we can now add more tests on top of this.

This also adds latest k8s versions to the test matrix.