Fix default_field for metricbeat (#7015)
In Elasticsearch 7.0 there is a limit of 1024 fields (see https://github.com/elastic/beats/issues/5275). As Metricbeat exceeds the limit of 1024 fields per index it has to be defined which the default fields are as otherwise Elasticsearch returns an error.

This change set the index config option `index.query.default_field`. To fields chosen to be searchable are all string and keyword values as these are the ones that I would expect users to put directly in the search without specifying a field. For all numbers I expect that the exact field is specified.

This PR also removes all presets from dashboards which did set default_fields as they were `*`. In Kibana the default `*` was removed also in master: https://github.com/elastic/kibana/issues/16232