Use `--mount` over `-v` syntax (#6725)
Reference https://docs.docker.com/storage/bind-mounts/

The `--mount` syntax is encouraged over the legacy `-v` syntax.

Also, eliminate the world "volume", because *a bind mount is not a volume*. A bind mount is a mount. A volume mount is a mount. A tmpfs mount is a mount (see https://docs.docker.com/storage/#choose-the-right-type-of-mount). These are the three types of mounts. Using the `--mount` syntax over the `-v` helps avoid this confusion.

Also, I changed the example to `$(pwd)` like they have it at https://docs.docker.com/storage/bind-mounts/#choosing-the--v-or---mount-flag because I was having trouble with `$(~)`, it didn't work for some reason (using bash)... Bind mounts require a fully qualified path (which is actually how the original example worked too because the `~` expands in the shell but in my changes the `~` would not expand because it is not the first character in the argument (not really sure how it works, but my example works))