Fetch github.com/docker/libcompose before testing libbeat
The libbeat tests were failing to compile and the error was uncaught. The cause was that github.com/docker/libcompose was not installed prior to running `go test` (relates to elastic/beats#5050). This moves `go get github.com/docker/libcompose` earlier in the script.

A separate change will be made in another commit to always report the `go test` exit code rather than the one from `go-junit-report`.