Log autodiscover config errors as Error instead of debug. (#8413)
If an invalid config is being generated, the user probably wants to know about it.

I personally found this to be quite confusing myself. Debug logging is usually for finding
either bugs in the code or extreme detail. I would say ensuring a config is valid is part
of the standard set of operations.