Make functionbeat build process depends on linux/amd64 (#8889)
Functionbeat packaging depends on a Linux binary that will be sent to
the serverless platform, previously the magefile for the project did not express
that dependency. This was causing problems when you were building the
packages only for a specific platform. This commit fixes that problem by
introducing `mage.WithPlatforms` this allows you to specific dependency
when you define the cross build logic.